+++ tests/testthat/test_downsample.R
@@ -41,7 +41,7 @@ test_that('basic usage', {
   expect_equal(sum(is.na(juice(rec1_p)$Species)), 5)
   expect_equal(te_xtab, og_xtab)
 
-  expect_warning(prep(rec1, training = iris2))
+  expect_warning(prep(rec1, training = iris2), NA)
 })
 
 test_that('ratio value', {